Cambio de float a double para el precio de las acciones

+++ StoXtreme/src/stoxtreme/cliente/GUI/MainFrameCliente.java
@@ -219,7 +219,7 @@ public class MainFrameCliente extends JFrame{
 							Operacion.COMPRA:Operacion.VENTA;
 				try{
 					int cantidad = Integer.parseInt(cantidadSeleccionada.getText());
-					float precio = Float.parseFloat(precioSeleccionado.getText());
+					double precio = Double.parseDouble(precioSeleccionado.getText());
 					new Operacion("Alonso", empresa, cantidad, precio, tipoOperacion);
 				}
 				catch(NumberFormatException ex){
